Add Spider and move median crawling to it
Spider is the new crawling application which crawls using periodic tasks
(as opposed to manually). After crawling the pages nightly, it checks
for changes and requests imports from the admin email as appropriate. It
also includes a dashboard for managing these pending imports.

This commit adds the Spider application and moves course median crawling
to it.